* upstream/master: (680 commits)
  Fix crash building Android rules
  Ignore classes within anonymous or local classes
  Fix annotations on inner class constructor params
  Enable `abi_verification_mode=fail`
  Fix stubs for deeply-nested static inner classes
  Fix references to deeply-nested enums
  Close cache at the end of cache command
  Add ndk.extra_ldflags
  Refactor creation of NDK ldflags
  Handle race conditions in Buck project constructor.
  ClassLoaderBootstrapper: Create ClassLoader earlier
  Stop logging "No Android platform target specified" to console
  JarBuilder: Eliminate checkNotNull
  Add more info for cannot delete directory failures
  AndroidLibraryDescription: Move ABI rule construction into the builder
  AndroidLibraryDescription: Move graph enhancement into the builder
  AndroidLibraryDescription: Push provided_deps_query into the builder
  DefaultJavaLibraryBuilder: Factor out getJavac
  DefaultJavaLibraryBuilder: Make initial BuildRuleParams accessible to subclasses
  Plumb TargetGraph and CellPathResolver into DefaultJavaLibraryBuilder
  ...

+++ programs/buck_tool.py
@@ -1,5 +1,6 @@
 from __future__ import print_function
 import errno
+import glob
 import json
 import os
 import platform
@@ -86,6 +87,11 @@ class CommandLineArgs:
     def is_help(self):
         return self.command is None or "--help" in self.command_options
 
+    # Whether this buck invocation is a normal buck or oop compilation mode
+    # TODO: remove this if oop javac is not good thing, or move to a truly independent jar
+    def is_oop_javac(self):
+        return self.command is None and "--oop-javac" in self.buck_options
+
 
 class RestartBuck(Exception):
     pass
@@ -95,8 +101,44 @@ class BuckToolException(Exception):
     pass
 
 
-class BuckTool(object):
+class JvmCrashLogger(object):
+    def __init__(self, buck_tool, project_root):
+        self._buck_tool = buck_tool
+        self._project_root = project_root
 
+    def __enter__(self):
+        self.jvm_errors_before = self._get_jvm_errors()
+
+    def __exit__(self, type, value, traceback):
+        new_errors = self._get_jvm_errors() - self.jvm_errors_before
+        if new_errors:
+            self._log_jvm_errors(new_errors)
+
+    def _get_jvm_errors(self):
+        return set(glob.glob(os.path.join(self._project_root, 'hs_err_pid*.log')))
+
+    @staticmethod
+    def _format_jvm_errors(fp):
+        errors = []
+        keep = False
+        for line in fp:
+            if line.startswith('Stack:'):
+                keep = True
+            if line.startswith('---------------  P R O C E S S  ---------------'):
+                break
+            if keep:
+                errors.append(line)
+        message = 'JVM native crash: ' + (errors[2] if len(errors) > 2 else '')
+        return message, ''.join(errors)
+
+    def _log_jvm_errors(self, errors):
+        for file in errors:
+            with open(file, 'r') as f:
+                message, loglines = self._format_jvm_errors(f)
+                print(loglines, file=sys.stderr)
+
+
+class BuckTool(object):
     def __init__(self, buck_project):
         self._init_timestamp = int(round(time.time() * 1000))
         self._command_line = CommandLineArgs(sys.argv)
@@ -143,7 +185,7 @@ class BuckTool(object):
 
     @property
     def _use_buckd(self):
-        return not os.environ.get('NO_BUCKD')
+        return not os.environ.get('NO_BUCKD') and not self._command_line.is_oop_javac()
 
     def _environ_for_buck(self):
         env = os.environ.copy()
@@ -154,84 +196,90 @@ class BuckTool(object):
 
     def launch_buck(self, build_id):
         with Tracing('BuckTool.launch_buck'):
-            if self._command_line.command == "clean" and not self._command_line.is_help():
-                self.kill_buckd()
+            with JvmCrashLogger(self, self._buck_project.root):
+                if self._command_line.command == "clean" and \
+                        not self._command_line.is_help() and \
+                        not self._command_line.is_oop_javac():
+                    self.kill_buckd()
 
-            buck_version_uid = self._get_buck_version_uid()
+                buck_version_uid = self._get_buck_version_uid()
 
-            use_buckd = self._use_buckd
-            if not self._command_line.is_help():
-                has_watchman = bool(which('watchman'))
-                if use_buckd and has_watchman:
-                    running_version = self._buck_project.get_running_buckd_version()
+                use_buckd = self._use_buckd
+                if not self._command_line.is_help() and not self._command_line.is_oop_javac():
+                    has_watchman = bool(which('watchman'))
+                    if use_buckd and has_watchman:
+                        running_version = self._buck_project.get_running_buckd_version()
 
-                    if running_version != buck_version_uid:
-                        self.kill_buckd()
+                        if running_version != buck_version_uid:
+                            self.kill_buckd()
 
-                    if not self._is_buckd_running():
-                        self.launch_buckd(buck_version_uid=buck_version_uid)
-                elif use_buckd and not has_watchman:
-                    print("Not using buckd because watchman isn't installed.",
-                          file=sys.stderr)
-                elif not use_buckd:
-                    print("Not using buckd because NO_BUCKD is set.",
-                          file=sys.stderr)
+                        if not self._is_buckd_running():
+                            self.launch_buckd(buck_version_uid=buck_version_uid)
+                    elif use_buckd and not has_watchman:
+                        print("Not using buckd because watchman isn't installed.",
+                              file=sys.stderr)
+                    elif not use_buckd:
+                        print("Not using buckd because NO_BUCKD is set.",
+                              file=sys.stderr)
 
-            env = self._environ_for_buck()
-            env['BUCK_BUILD_ID'] = build_id
+                env = self._environ_for_buck()
+                env['BUCK_BUILD_ID'] = build_id
 
-            buck_socket_path = self._buck_project.get_buckd_socket_path()
+                if use_buckd and self._is_buckd_running():
+                    with Tracing('buck', args={'command': sys.argv[1:]}):
+                        exit_code = 2
+                        last_diagnostic_time = 0
+                        while exit_code == 2:
+                            with NailgunConnection(
+                                    self._buck_project.get_buckd_transport_address(),
+                                    cwd=self._buck_project.root) as c:
+                                now = int(round(time.time() * 1000))
+                                env['BUCK_PYTHON_SPACE_INIT_TIME'] = \
+                                    str(now - self._init_timestamp)
+                                exit_code = c.send_command(
+                                    'com.facebook.buck.cli.Main',
+                                    sys.argv[1:],
+                                    env=env,
+                                    cwd=self._buck_project.root)
+                                if exit_code == 2:
+                                    env['BUCK_BUILD_ID'] = str(uuid.uuid4())
+                                    now = time.time()
+                                    if now - last_diagnostic_time > DAEMON_BUSY_MESSAGE_SECONDS:
+                                        print('Daemon is busy, waiting for it to become free...',
+                                              file=sys.stderr)
+                                        last_diagnostic_time = now
+                                    time.sleep(1)
+                        return exit_code
 
-            if use_buckd and self._is_buckd_running() and \
-                    os.path.exists(buck_socket_path):
-                with Tracing('buck', args={'command': sys.argv[1:]}):
-                    exit_code = 2
-                    last_diagnostic_time = 0
-                    while exit_code == 2:
-                        with NailgunConnection('local:.buckd/sock',
-                                               cwd=self._buck_project.root) as c:
-                            now = int(round(time.time() * 1000))
-                            env['BUCK_PYTHON_SPACE_INIT_TIME'] = str(now - self._init_timestamp)
-                            exit_code = c.send_command(
-                                'com.facebook.buck.cli.Main',
-                                sys.argv[1:],
-                                env=env,
-                                cwd=self._buck_project.root)
-                            if exit_code == 2:
-                                env['BUCK_BUILD_ID'] = str(uuid.uuid4())
-                                now = time.time()
-                                if now - last_diagnostic_time > DAEMON_BUSY_MESSAGE_SECONDS:
-                                    print('Daemon is busy, waiting for it to become free...',
-                                          file=sys.stderr)
-                                    last_diagnostic_time = now
-                                time.sleep(1)
-                    return exit_code
+                command = ["buck"]
+                extra_default_options = [
+                    "-Djava.io.tmpdir={0}".format(self._tmp_dir),
+                    "-XX:SoftRefLRUPolicyMSPerMB=0",
+                    "-XX:+UseG1GC",
+                ]
+                command.extend(self._get_java_args(buck_version_uid, extra_default_options))
+                command.append("com.facebook.buck.cli.bootstrapper.ClassLoaderBootstrapper")
+                if self._command_line.is_oop_javac():
+                    command.append("com.facebook.buck.oop_javac.Main")
+                else:
+                    command.append("com.facebook.buck.cli.Main")
+                command.extend(sys.argv[1:])
+
+                now = int(round(time.time() * 1000))
+                env['BUCK_PYTHON_SPACE_INIT_TIME'] = str(now - self._init_timestamp)
+                if True:
+                    java = which("java")
+                    if java is None:
+                        raise BuckToolException('Could not find java on $PATH')
+                    with Tracing('buck', args={'command': command}):
+                        buck_exit_code = subprocess.call(command,
+                                                         cwd=self._buck_project.root,
+                                                         env=env,
+                                                         executable=java)
+                return buck_exit_code
 
 
-            command = ["buck"]
-            extra_default_options = [
-                "-Djava.io.tmpdir={0}".format(self._tmp_dir)
-            ]
-            command.extend(self._get_java_args(buck_version_uid, extra_default_options))
-            command.append("com.facebook.buck.cli.bootstrapper.ClassLoaderBootstrapper")
-            command.append("com.facebook.buck.cli.Main")
-            command.extend(sys.argv[1:])
-
-            now = int(round(time.time() * 1000))
-            env['BUCK_PYTHON_SPACE_INIT_TIME'] = str(now - self._init_timestamp)
-            if True:
-                java = which("java")
-                if java is None:
-                    raise BuckToolException('Could not find java on $PATH')
-                with Tracing('buck', args={'command': command}):
-                    buck_exit_code = subprocess.call(command,
-                                                     cwd=self._buck_project.root,
-                                                     env=env,
-                                                     executable=java)
-            return buck_exit_code
-
-
-    def _generate_log_entry(self, exit_code, message, logs_array):
+    def _generate_log_entry(self, message, logs_array):
         import socket
         import getpass
         traits = {
@@ -246,12 +294,12 @@ class BuckTool(object):
             "isDaemon": "false",
         }
         entry = {
-          "logs": logs_array,
-          "traits": traits,
-          "message": message,
-          "category": message,
-          "time": int(time.time()),
-          "logger": "com.facebook.buck.python.buck_tool.py",
+            "logs": logs_array,
+            "traits": traits,
+            "message": message,
+            "category": message,
+            "time": int(time.time()),
+            "logger": "com.facebook.buck.python.buck_tool.py",
         }
         return entry
 
@@ -297,47 +345,55 @@ class BuckTool(object):
             command.extend(self._get_java_args(buck_version_uid, extra_default_options))
             command.append("com.facebook.buck.cli.bootstrapper.ClassLoaderBootstrapper")
             command.append("com.facebook.buck.cli.Main$DaemonBootstrap")
-            command.append("local:.buckd/sock")
+            command.append(self._buck_project.get_buckd_transport_address())
             command.append("{0}".format(BUCKD_CLIENT_TIMEOUT_MILLIS))
 
-            '''
-            Change the process group of the child buckd process so that when this
-            script is interrupted, it does not kill buckd.
-            '''
-            def preexec_func():
-                # Close any open file descriptors to further separate buckd from its
-                # invoking context (e.g. otherwise we'd hang when running things like
-                # `ssh localhost buck clean`).
-                dev_null_fd = os.open("/dev/null", os.O_RDWR)
-                os.dup2(dev_null_fd, 0)
-                os.dup2(dev_null_fd, 1)
-                os.dup2(dev_null_fd, 2)
-                os.close(dev_null_fd)
-            buck_socket_path = self._buck_project.get_buckd_socket_path()
+            buckd_transport_file_path = self._buck_project.get_buckd_transport_file_path()
+            if os.name == 'nt':
+                preexec_fn = None
+                # https://msdn.microsoft.com/en-us/library/windows/desktop/ms684863.aspx#DETACHED_PROCESS
+                DETACHED_PROCESS = 0x00000008
+                creationflags = DETACHED_PROCESS
+            else:
+                # Make sure the Unix domain socket doesn't exist before this call.
+                try:
+                    os.unlink(buckd_transport_file_path)
+                except OSError as e:
+                    if e.errno == errno.ENOENT:
+                        # Socket didn't previously exist.
+                        pass
+                    else:
+                        raise e
+                '''
+                Change the process group of the child buckd process so that when this
+                script is interrupted, it does not kill buckd.
+                '''
 
-            # Make sure the Unix domain socket doesn't exist before this call.
-            try:
-                os.unlink(buck_socket_path)
-            except OSError as e:
-                if e.errno == errno.ENOENT:
-                    # Socket didn't previously exist.
-                    pass
-                else:
-                    raise e
+                def preexec_fn():
+                    # Close any open file descriptors to further separate buckd from its
+                    # invoking context (e.g. otherwise we'd hang when running things like
+                    # `ssh localhost buck clean`).
+                    dev_null_fd = os.open("/dev/null", os.O_RDWR)
+                    os.dup2(dev_null_fd, 0)
+                    os.dup2(dev_null_fd, 1)
+                    os.dup2(dev_null_fd, 2)
+                    os.close(dev_null_fd)
 
+                creationflags = 0
             process = subprocess.Popen(
                 command,
                 executable=which("java"),
                 cwd=self._buck_project.root,
                 close_fds=True,
-                preexec_fn=preexec_func,
-                env=self._environ_for_buck())
+                preexec_fn=preexec_fn,
+                env=self._environ_for_buck(),
+                creationflags=creationflags)
 
             self._buck_project.save_buckd_version(buck_version_uid)
 
             # Give Java some time to create the listening socket.
             for i in range(0, 300):
-                if not os.path.exists(buck_socket_path):
+                if not transport_exists(buckd_transport_file_path):
                     time.sleep(0.01)
 
             returncode = process.poll()
@@ -350,11 +406,12 @@ class BuckTool(object):
 
     def kill_buckd(self):
         with Tracing('BuckTool.kill_buckd'):
-            buckd_socket_path = self._buck_project.get_buckd_socket_path()
-            if os.path.exists(buckd_socket_path):
+            buckd_transport_file_path = self._buck_project.get_buckd_transport_file_path()
+            if transport_exists(buckd_transport_file_path):
                 print("Shutting down nailgun server...", file=sys.stderr)
                 try:
-                    with NailgunConnection('local:.buckd/sock', cwd=self._buck_project.root) as c:
+                    with NailgunConnection(self._buck_project.get_buckd_transport_address(),
+                                           cwd=self._buck_project.root) as c:
                         c.send_command('ng-stop')
                 except NailgunException as e:
                     if e.code not in (NailgunException.CONNECT_FAILED,
@@ -368,14 +425,14 @@ class BuckTool(object):
 
     def _is_buckd_running(self):
         with Tracing('BuckTool._is_buckd_running'):
-            buckd_socket_path = self._buck_project.get_buckd_socket_path()
+            transport_file_path = self._buck_project.get_buckd_transport_file_path()
 
-            if not os.path.exists(buckd_socket_path):
+            if not transport_exists(transport_file_path):
                 return False
 
             try:
                 with NailgunConnection(
-                        'local:.buckd/sock',
+                        self._buck_project.get_buckd_transport_address(),
                         stdin=None,
                         stdout=None,
                         stderr=None,
@@ -476,3 +533,26 @@ def setup_watchman_watch():
             raise BuckToolException(message)
 
         print("Using watchman.", file=sys.stderr)
+
+
+def transport_exists(path):
+    return os.path.exists(path)
+
+
+if os.name == 'nt':
+    import ctypes
+    from ctypes.wintypes import WIN32_FIND_DATAW as WIN32_FIND_DATA
+
+    INVALID_HANDLE_VALUE = -1
+    FindFirstFile = ctypes.windll.kernel32.FindFirstFileW
+    FindClose = ctypes.windll.kernel32.FindClose
+
+
+    # on windows os.path.exists doen't allow to check reliably that pipe exists
+    # (os.path.exists tries to open connection to a pipe)
+    def transport_exists(transport_path):
+        wfd = WIN32_FIND_DATA()
+        handle = FindFirstFile(transport_path, ctypes.byref(wfd))
+        result = handle != INVALID_HANDLE_VALUE
+        FindClose(handle)
+        return result
